package location refactors
