+++ docs/allclasses-frame.html
@@ -78,8 +78,8 @@
 <li><a href="ure/UAnimator.html" title="interface in ure" target="classFrame"><span class="interfaceName">UAnimator</span></a></li>
 <li><a href="ure/areas/UArea.html" title="class in ure.areas" target="classFrame">UArea</a></li>
 <li><a href="ure/areas/UArea.Listener.html" title="interface in ure.areas" target="classFrame"><span class="interfaceName">UArea.Listener</span></a></li>
-<li><a href="ure/behaviors/UBehavior.html" title="class in ure.behaviors" target="classFrame">UBehavior</a></li>
-<li><a href="ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors" target="classFrame">UBehaviorWander</a></li>
+<li><a href="ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors" target="classFrame">UBehavior</a></li>
+<li><a href="ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors" target="classFrame">UBehaviorWander</a></li>
 <li><a href="ure/ui/UCamera.html" title="class in ure.ui" target="classFrame">UCamera</a></li>
 <li><a href="ure/areas/UCartographer.html" title="class in ure.areas" target="classFrame">UCartographer</a></li>
 <li><a href="ure/areas/UCell.html" title="class in ure.areas" target="classFrame">UCell</a></li>
