+++ docs/allclasses-noframe.html
@@ -78,8 +78,8 @@
 <li><a href="ure/UAnimator.html" title="interface in ure"><span class="interfaceName">UAnimator</span></a></li>
 <li><a href="ure/areas/UArea.html" title="class in ure.areas">UArea</a></li>
 <li><a href="ure/areas/UArea.Listener.html" title="interface in ure.areas"><span class="interfaceName">UArea.Listener</span></a></li>
-<li><a href="ure/behaviors/UBehavior.html" title="class in ure.behaviors">UBehavior</a></li>
-<li><a href="ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors">UBehaviorWander</a></li>
+<li><a href="ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors">UBehavior</a></li>
+<li><a href="ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors">UBehaviorWander</a></li>
 <li><a href="ure/ui/UCamera.html" title="class in ure.ui">UCamera</a></li>
 <li><a href="ure/areas/UCartographer.html" title="class in ure.areas">UCartographer</a></li>
 <li><a href="ure/areas/UCell.html" title="class in ure.areas">UCell</a></li>
