+++ docs/element-list
@@ -2,7 +2,7 @@ ure
 ure.actions
 ure.actors
 ure.areas
-ure.behaviors
+ure.actors.behaviors
 ure.commands
 ure.sys.dagger
 ure.examplegame
