+++ docs/index-all.html
@@ -112,11 +112,11 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="ure/actors/UNPC.html#act--">act()</a></span> - Method in class ure.actors.<a href="ure/actors/UNPC.html" title="class in ure.actors">UNPC</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="ure/behaviors/UBehavior.html#action-ure.actors.UNPC-">action(UNPC)</a></span> - Method in class ure.behaviors.<a href="ure/behaviors/UBehavior.html" title="class in ure.behaviors">UBehavior</a></dt>
+<dt><span class="memberNameLink"><a href="ure/behaviors/UBehavior.html#action-ure.actors.UNPC-">action(UNPC)</a></span> - Method in class ure.actors.behaviors.<a href="ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors">UBehavior</a></dt>
 <dd>
 <div class="block">Override action() to return UActions on demand for the actor to perform.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="ure/behaviors/UBehaviorWander.html#action-ure.actors.UNPC-">action(UNPC)</a></span> - Method in class ure.behaviors.<a href="ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors">UBehaviorWander</a></dt>
+<dt><span class="memberNameLink"><a href="ure/behaviors/UBehaviorWander.html#action-ure.actors.UNPC-">action(UNPC)</a></span> - Method in class ure.actors.behaviors.<a href="ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors">UBehaviorWander</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="ure/actors/UActor.html#actionSpeed--">actionSpeed()</a></span> - Method in class ure.actors.<a href="ure/actors/UActor.html" title="class in ure.actors">UActor</a></dt>
 <dd>&nbsp;</dd>
@@ -1602,9 +1602,9 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="ure/actors/Blank.html#TYPE">TYPE</a></span> - Static variable in class ure.actors.<a href="ure/actors/Blank.html" title="class in ure.actors">Blank</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="ure/behaviors/UBehavior.html#TYPE">TYPE</a></span> - Static variable in class ure.behaviors.<a href="ure/behaviors/UBehavior.html" title="class in ure.behaviors">UBehavior</a></dt>
+<dt><span class="memberNameLink"><a href="ure/behaviors/UBehavior.html#TYPE">TYPE</a></span> - Static variable in class ure.actors.behaviors.<a href="ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors">UBehavior</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="ure/behaviors/UBehaviorWander.html#TYPE">TYPE</a></span> - Static variable in class ure.behaviors.<a href="ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors">UBehaviorWander</a></dt>
+<dt><span class="memberNameLink"><a href="ure/behaviors/UBehaviorWander.html#TYPE">TYPE</a></span> - Static variable in class ure.actors.behaviors.<a href="ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors">UBehaviorWander</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="ure/terrain/Door.html#TYPE">TYPE</a></span> - Static variable in class ure.terrain.<a href="ure/terrain/Door.html" title="class in ure.terrain">Door</a></dt>
 <dd>&nbsp;</dd>
@@ -1704,18 +1704,18 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>&nbsp;</dd>
 <dt><a href="ure/areas/UArea.Listener.html" title="interface in ure.areas"><span class="typeNameLink">UArea.Listener</span></a> - Interface in <a href="ure/areas/package-summary.html">ure.areas</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="ure/behaviors/UBehavior.html" title="class in ure.behaviors"><span class="typeNameLink">UBehavior</span></a> - Class in <a href="ure/behaviors/package-summary.html">ure.behaviors</a></dt>
+<dt><a href="ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors"><span class="typeNameLink">UBehavior</span></a> - Class in <a href="ure/behaviors/package-summary.html">ure.actors.behaviors</a></dt>
 <dd>
 <div class="block">UBehavior implements a source of actions for an NPC actor to perform, to emulate a behavior or achieve
  a goal.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="ure/behaviors/UBehavior.html#UBehavior--">UBehavior()</a></span> - Constructor for class ure.behaviors.<a href="ure/behaviors/UBehavior.html" title="class in ure.behaviors">UBehavior</a></dt>
+<dt><span class="memberNameLink"><a href="ure/behaviors/UBehavior.html#UBehavior--">UBehavior()</a></span> - Constructor for class ure.actors.behaviors.<a href="ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors">UBehavior</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors"><span class="typeNameLink">UBehaviorWander</span></a> - Class in <a href="ure/behaviors/package-summary.html">ure.behaviors</a></dt>
+<dt><a href="ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors"><span class="typeNameLink">UBehaviorWander</span></a> - Class in <a href="ure/behaviors/package-summary.html">ure.actors.behaviors</a></dt>
 <dd>
 <div class="block">Wander aimlessly.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="ure/behaviors/UBehaviorWander.html#UBehaviorWander--">UBehaviorWander()</a></span> - Constructor for class ure.behaviors.<a href="ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors">UBehaviorWander</a></dt>
+<dt><span class="memberNameLink"><a href="ure/behaviors/UBehaviorWander.html#UBehaviorWander--">UBehaviorWander()</a></span> - Constructor for class ure.actors.behaviors.<a href="ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors">UBehaviorWander</a></dt>
 <dd>&nbsp;</dd>
 <dt><a href="ure/ui/UCamera.html" title="class in ure.ui"><span class="typeNameLink">UCamera</span></a> - Class in <a href="ure/ui/package-summary.html">ure.ui</a></dt>
 <dd>
@@ -1829,7 +1829,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>
 <div class="block">Areas are self contained contiguous 2D grids of Cells, representing a game playfield.</div>
 </dd>
-<dt><a href="ure/behaviors/package-summary.html">ure.behaviors</a> - package ure.behaviors</dt>
+<dt><a href="ure/behaviors/package-summary.html">ure.actors.behaviors</a> - package ure.actors.behaviors</dt>
 <dd>
 <div class="block">Behaviors are members of NPCs which emit streams of Actions.</div>
 </dd>
