+++ docs/index-files/index-1.html
@@ -112,11 +112,11 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../ure/actors/UNPC.html#act--">act()</a></span> - Method in class ure.actors.<a href="../ure/actors/UNPC.html" title="class in ure.actors">UNPC</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../ure/behaviors/UBehavior.html#action-ure.actors.UNPC-">action(UNPC)</a></span> - Method in class ure.behaviors.<a href="../ure/behaviors/UBehavior.html" title="class in ure.behaviors">UBehavior</a></dt>
+<dt><span class="memberNameLink"><a href="../ure/behaviors/UBehavior.html#action-ure.actors.UNPC-">action(UNPC)</a></span> - Method in class ure.actors.behaviors.<a href="../ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors">UBehavior</a></dt>
 <dd>
 <div class="block">Override action() to return UActions on demand for the actor to perform.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../ure/behaviors/UBehaviorWander.html#action-ure.actors.UNPC-">action(UNPC)</a></span> - Method in class ure.behaviors.<a href="../ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors">UBehaviorWander</a></dt>
+<dt><span class="memberNameLink"><a href="../ure/behaviors/UBehaviorWander.html#action-ure.actors.UNPC-">action(UNPC)</a></span> - Method in class ure.actors.behaviors.<a href="../ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors">UBehaviorWander</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../ure/actors/UActor.html#actionSpeed--">actionSpeed()</a></span> - Method in class ure.actors.<a href="../ure/actors/UActor.html" title="class in ure.actors">UActor</a></dt>
 <dd>&nbsp;</dd>
