+++ docs/index-files/index-18.html
@@ -192,9 +192,9 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../ure/actors/Blank.html#TYPE">TYPE</a></span> - Static variable in class ure.actors.<a href="../ure/actors/Blank.html" title="class in ure.actors">Blank</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../ure/behaviors/UBehavior.html#TYPE">TYPE</a></span> - Static variable in class ure.behaviors.<a href="../ure/behaviors/UBehavior.html" title="class in ure.behaviors">UBehavior</a></dt>
+<dt><span class="memberNameLink"><a href="../ure/behaviors/UBehavior.html#TYPE">TYPE</a></span> - Static variable in class ure.actors.behaviors.<a href="../ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors">UBehavior</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../ure/behaviors/UBehaviorWander.html#TYPE">TYPE</a></span> - Static variable in class ure.behaviors.<a href="../ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors">UBehaviorWander</a></dt>
+<dt><span class="memberNameLink"><a href="../ure/behaviors/UBehaviorWander.html#TYPE">TYPE</a></span> - Static variable in class ure.actors.behaviors.<a href="../ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors">UBehaviorWander</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../ure/terrain/Door.html#TYPE">TYPE</a></span> - Static variable in class ure.terrain.<a href="../ure/terrain/Door.html" title="class in ure.terrain">Door</a></dt>
 <dd>&nbsp;</dd>
