+++ docs/index-files/index-19.html
@@ -172,18 +172,18 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>&nbsp;</dd>
 <dt><a href="../ure/areas/UArea.Listener.html" title="interface in ure.areas"><span class="typeNameLink">UArea.Listener</span></a> - Interface in <a href="../ure/areas/package-summary.html">ure.areas</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../ure/behaviors/UBehavior.html" title="class in ure.behaviors"><span class="typeNameLink">UBehavior</span></a> - Class in <a href="../ure/behaviors/package-summary.html">ure.behaviors</a></dt>
+<dt><a href="../ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors"><span class="typeNameLink">UBehavior</span></a> - Class in <a href="../ure/behaviors/package-summary.html">ure.actors.behaviors</a></dt>
 <dd>
 <div class="block">UBehavior implements a source of actions for an NPC actor to perform, to emulate a behavior or achieve
  a goal.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../ure/behaviors/UBehavior.html#UBehavior--">UBehavior()</a></span> - Constructor for class ure.behaviors.<a href="../ure/behaviors/UBehavior.html" title="class in ure.behaviors">UBehavior</a></dt>
+<dt><span class="memberNameLink"><a href="../ure/behaviors/UBehavior.html#UBehavior--">UBehavior()</a></span> - Constructor for class ure.actors.behaviors.<a href="../ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors">UBehavior</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors"><span class="typeNameLink">UBehaviorWander</span></a> - Class in <a href="../ure/behaviors/package-summary.html">ure.behaviors</a></dt>
+<dt><a href="../ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors"><span class="typeNameLink">UBehaviorWander</span></a> - Class in <a href="../ure/behaviors/package-summary.html">ure.actors.behaviors</a></dt>
 <dd>
 <div class="block">Wander aimlessly.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../ure/behaviors/UBehaviorWander.html#UBehaviorWander--">UBehaviorWander()</a></span> - Constructor for class ure.behaviors.<a href="../ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors">UBehaviorWander</a></dt>
+<dt><span class="memberNameLink"><a href="../ure/behaviors/UBehaviorWander.html#UBehaviorWander--">UBehaviorWander()</a></span> - Constructor for class ure.actors.behaviors.<a href="../ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors">UBehaviorWander</a></dt>
 <dd>&nbsp;</dd>
 <dt><a href="../ure/ui/UCamera.html" title="class in ure.ui"><span class="typeNameLink">UCamera</span></a> - Class in <a href="../ure/ui/package-summary.html">ure.ui</a></dt>
 <dd>
@@ -313,7 +313,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>
 <div class="block">Areas are self contained contiguous 2D grids of Cells, representing a game playfield.</div>
 </dd>
-<dt><a href="../ure/behaviors/package-summary.html">ure.behaviors</a> - package ure.behaviors</dt>
+<dt><a href="../ure/behaviors/package-summary.html">ure.actors.behaviors</a> - package ure.actors.behaviors</dt>
 <dd>
 <div class="block">Behaviors are members of NPCs which emit streams of Actions.</div>
 </dd>
