+++ docs/index-files/index-9.html
@@ -200,7 +200,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../ure/dagger/AppComponent.html#inject-ure.areas.UCell-">inject(UCell)</a></span> - Method in interface ure.sys.dagger.<a href="../ure/dagger/AppComponent.html" title="interface in ure.sys.dagger">AppComponent</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../ure/dagger/AppComponent.html#inject-ure.behaviors.UBehavior-">inject(UBehavior)</a></span> - Method in interface ure.sys.dagger.<a href="../ure/dagger/AppComponent.html" title="interface in ure.sys.dagger">AppComponent</a></dt>
+<dt><span class="memberNameLink"><a href="../ure/dagger/AppComponent.html#inject-ure.actors.behaviors.UBehavior-">inject(UBehavior)</a></span> - Method in interface ure.sys.dagger.<a href="../ure/dagger/AppComponent.html" title="interface in ure.sys.dagger">AppComponent</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../ure/dagger/AppComponent.html#inject-ure.commands.UCommand-">inject(UCommand)</a></span> - Method in interface ure.sys.dagger.<a href="../ure/dagger/AppComponent.html" title="interface in ure.sys.dagger">AppComponent</a></dt>
 <dd>&nbsp;</dd>
