+++ docs/overview-frame.html
@@ -30,7 +30,7 @@
 <li><a href="ure/actions/package-frame.html" target="packageFrame">ure.actions</a></li>
 <li><a href="ure/actors/package-frame.html" target="packageFrame">ure.actors</a></li>
 <li><a href="ure/areas/package-frame.html" target="packageFrame">ure.areas</a></li>
-<li><a href="ure/behaviors/package-frame.html" target="packageFrame">ure.behaviors</a></li>
+<li><a href="ure/behaviors/package-frame.html" target="packageFrame">ure.actors.behaviors</a></li>
 <li><a href="ure/commands/package-frame.html" target="packageFrame">ure.commands</a></li>
 <li><a href="ure/dagger/package-frame.html" target="packageFrame">ure.sys.dagger</a></li>
 <li><a href="ure/examplegame/package-frame.html" target="packageFrame">ure.examplegame</a></li>
