+++ docs/overview-summary.html
@@ -127,7 +127,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </td>
 </tr>
 <tr class="altColor" id="i4">
-<th class="colFirst" scope="row"><a href="ure/behaviors/package-summary.html">ure.behaviors</a></th>
+<th class="colFirst" scope="row"><a href="ure/behaviors/package-summary.html">ure.actors.behaviors</a></th>
 <td class="colLast">
 <div class="block">Behaviors are members of NPCs which emit streams of Actions.</div>
 </td>
