+++ docs/overview-tree.html
@@ -101,7 +101,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li><a href="ure/actions/package-tree.html">ure.actions</a>, </li>
 <li><a href="ure/actors/package-tree.html">ure.actors</a>, </li>
 <li><a href="ure/areas/package-tree.html">ure.areas</a>, </li>
-<li><a href="ure/behaviors/package-tree.html">ure.behaviors</a>, </li>
+<li><a href="ure/behaviors/package-tree.html">ure.actors.behaviors</a>, </li>
 <li><a href="ure/commands/package-tree.html">ure.commands</a>, </li>
 <li><a href="ure/dagger/package-tree.html">ure.sys.dagger</a>, </li>
 <li><a href="ure/examplegame/package-tree.html">ure.examplegame</a>, </li>
@@ -174,9 +174,9 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </li>
 <li class="circle">ure.actors.<a href="ure/actors/UActorCzar.html" title="class in ure.actors"><span class="typeNameLink">UActorCzar</span></a></li>
 <li class="circle">ure.areas.<a href="ure/areas/UArea.html" title="class in ure.areas"><span class="typeNameLink">UArea</span></a> (implements java.io.Serializable, ure.<a href="ure/UTimeListener.html" title="interface in ure">UTimeListener</a>)</li>
-<li class="circle">ure.behaviors.<a href="ure/behaviors/UBehavior.html" title="class in ure.behaviors"><span class="typeNameLink">UBehavior</span></a>
+<li class="circle">ure.actors.behaviors.<a href="ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors"><span class="typeNameLink">UBehavior</span></a>
 <ul>
-<li class="circle">ure.behaviors.<a href="ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors"><span class="typeNameLink">UBehaviorWander</span></a></li>
+<li class="circle">ure.actors.behaviors.<a href="ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors"><span class="typeNameLink">UBehaviorWander</span></a></li>
 </ul>
 </li>
 <li class="circle">ure.areas.<a href="ure/areas/UCartographer.html" title="class in ure.areas"><span class="typeNameLink">UCartographer</span></a></li>
