+++ docs/ure/behaviors/UBehavior.html
@@ -59,7 +59,7 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 <div class="subNav">
 <ul class="navList">
 <li>Prev&nbsp;Class</li>
-<li><a href="../../ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
 <li><a href="../../index.html?ure/behaviors/UBehavior.html" target="_top">Frames</a></li>
@@ -116,7 +116,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </script>
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../ure/behaviors/package-summary.html">ure.behaviors</a></div>
+<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../ure/behaviors/package-summary.html">ure.actors.behaviors</a></div>
 <h2 title="Class UBehavior" class="title">Class UBehavior</h2>
 </div>
 <div class="contentContainer">
@@ -124,7 +124,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>ure.behaviors.UBehavior</li>
+<li>ure.actors.behaviors.UBehavior</li>
 </ul>
 </li>
 </ul>
@@ -133,7 +133,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li class="blockList">
 <dl>
 <dt>Direct Known Subclasses:</dt>
-<dd><code><a href="../../ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors">UBehaviorWander</a></code></dd>
+<dd><code><a href="../../ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors">UBehaviorWander</a></code></dd>
 </dl>
 <hr>
 <pre>public abstract class <span class="typeNameLabel">UBehavior</span>
@@ -308,7 +308,7 @@ extends java.lang.Object</pre>
 <div class="subNav">
 <ul class="navList">
 <li>Prev&nbsp;Class</li>
-<li><a href="../../ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
 <li><a href="../../index.html?ure/behaviors/UBehavior.html" target="_top">Frames</a></li>
