+++ docs/ure/behaviors/UBehaviorWander.html
@@ -58,7 +58,7 @@ var pathtoroot = "../../";loadScripts(document, 'script');</script>
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../ure/behaviors/UBehavior.html" title="class in ure.behaviors"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
@@ -116,7 +116,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </script>
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../ure/behaviors/package-summary.html">ure.behaviors</a></div>
+<div class="subTitle"><span class="packageLabelInType">Package</span>&nbsp;<a href="../../ure/behaviors/package-summary.html">ure.actors.behaviors</a></div>
 <h2 title="Class UBehaviorWander" class="title">Class UBehaviorWander</h2>
 </div>
 <div class="contentContainer">
@@ -124,10 +124,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li><a href="../../ure/behaviors/UBehavior.html" title="class in ure.behaviors">ure.behaviors.UBehavior</a></li>
+<li><a href="../../ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors">ure.actors.behaviors.UBehavior</a></li>
 <li>
 <ul class="inheritance">
-<li>ure.behaviors.BehaviorWander</li>
+<li>ure.actors.behaviors.BehaviorWander</li>
 </ul>
 </li>
 </ul>
@@ -138,7 +138,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li class="blockList">
 <hr>
 <pre>public class <span class="typeNameLabel">UBehaviorWander</span>
-extends <a href="../../ure/behaviors/UBehavior.html" title="class in ure.behaviors">UBehavior</a></pre>
+extends <a href="../../ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors">UBehavior</a></pre>
 <div class="block">Wander aimlessly.</div>
 </li>
 </ul>
@@ -273,7 +273,7 @@ extends <a href="../../ure/behaviors/UBehavior.html" title="class in ure.behavio
 <div class="block">Override action() to return UActions on demand for the actor to perform.</div>
 <dl>
 <dt><span class="overrideSpecifyLabel">Overrides:</span></dt>
-<dd><code><a href="../../ure/behaviors/UBehavior.html#action-ure.actors.UNPC-">action</a></code>&nbsp;in class&nbsp;<code><a href="../../ure/behaviors/UBehavior.html" title="class in ure.behaviors">UBehavior</a></code></dd>
+<dd><code><a href="../../ure/behaviors/UBehavior.html#action-ure.actors.UNPC-">action</a></code>&nbsp;in class&nbsp;<code><a href="../../ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors">UBehavior</a></code></dd>
 <dt><span class="returnLabel">Returns:</span></dt>
 </dl>
 </li>
@@ -305,7 +305,7 @@ extends <a href="../../ure/behaviors/UBehavior.html" title="class in ure.behavio
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../ure/behaviors/UBehavior.html" title="class in ure.behaviors"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
