+++ docs/ure/behaviors/package-frame.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (10.0.1) on Wed Jul 11 18:21:26 PDT 2018 -->
-<title>ure.behaviors</title>
+<title>ure.actors.behaviors</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="date" content="2018-07-11">
 <link rel="stylesheet" type="text/css" href="../../stylesheet.css" title="Style">
@@ -18,12 +18,12 @@
 <script type="text/javascript" src="../../jquery/jquery-ui.js"></script>
 </head>
 <body>
-<h1 class="bar"><a href="../../ure/behaviors/package-summary.html" target="classFrame">ure.behaviors</a></h1>
+<h1 class="bar"><a href="../../ure/behaviors/package-summary.html" target="classFrame">ure.actors.behaviors</a></h1>
 <div class="indexContainer">
 <h2 title="Classes">Classes</h2>
 <ul title="Classes">
-<li><a href="UBehavior.html" title="class in ure.behaviors" target="classFrame">UBehavior</a></li>
-<li><a href="UBehaviorWander.html" title="class in ure.behaviors" target="classFrame">UBehaviorWander</a></li>
+<li><a href="UBehavior.html" title="class in ure.actors.behaviors" target="classFrame">UBehavior</a></li>
+<li><a href="UBehaviorWander.html" title="class in ure.actors.behaviors" target="classFrame">UBehaviorWander</a></li>
 </ul>
 </div>
 </body>
