+++ docs/ure/behaviors/package-summary.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (10.0.1) on Wed Jul 11 18:21:26 PDT 2018 -->
-<title>ure.behaviors</title>
+<title>ure.actors.behaviors</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="date" content="2018-07-11">
 <link rel="stylesheet" type="text/css" href="../../stylesheet.css" title="Style">
@@ -21,7 +21,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="ure.behaviors";
+            parent.document.title="ure.actors.behaviors";
         }
     }
     catch(err) {
@@ -94,7 +94,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 //-->
 </script>
 <div class="header">
-<h1 title="Package" class="title">Package&nbsp;ure.behaviors</h1>
+<h1 title="Package" class="title">Package&nbsp;ure.actors.behaviors</h1>
 </div>
 <div class="contentContainer"><a name="package.description">
 <!--   -->
@@ -110,14 +110,14 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </tr>
 <tbody>
 <tr class="altColor">
-<th class="colFirst" scope="row"><a href="../../ure/behaviors/UBehavior.html" title="class in ure.behaviors">UBehavior</a></th>
+<th class="colFirst" scope="row"><a href="../../ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors">UBehavior</a></th>
 <td class="colLast">
 <div class="block">UBehavior implements a source of actions for an NPC actor to perform, to emulate a behavior or achieve
  a goal.</div>
 </td>
 </tr>
 <tr class="rowColor">
-<th class="colFirst" scope="row"><a href="../../ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors">UBehaviorWander</a></th>
+<th class="colFirst" scope="row"><a href="../../ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors">UBehaviorWander</a></th>
 <td class="colLast">
 <div class="block">Wander aimlessly.</div>
 </td>
