+++ docs/ure/behaviors/package-tree.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (10.0.1) on Wed Jul 11 18:21:26 PDT 2018 -->
-<title>ure.behaviors Class Hierarchy</title>
+<title>ure.actors.behaviors Class Hierarchy</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="date" content="2018-07-11">
 <link rel="stylesheet" type="text/css" href="../../stylesheet.css" title="Style">
@@ -21,13 +21,12 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="ure.behaviors Class Hierarchy";
+            parent.document.title="ure.actors.behaviors Class Hierarchy";
         }
     }
     catch(err) {
     }
-//-->
-var pathtoroot = "../../";loadScripts(document, 'script');</script>
+//--></script>
 <noscript>
 <div>JavaScript is disabled on your browser.</div>
 </noscript>
@@ -94,7 +93,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 //-->
 </script>
 <div class="header">
-<h1 class="title">Hierarchy For Package ure.behaviors</h1>
+<h1 class="title">Hierarchy For Package ure.actors.behaviors</h1>
 <span class="packageHierarchyLabel">Package Hierarchies:</span>
 <ul class="horizontal">
 <li><a href="../../overview-tree.html">All Packages</a></li>
@@ -105,9 +104,9 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <ul>
 <li class="circle">java.lang.Object
 <ul>
-<li class="circle">ure.behaviors.<a href="../../ure/behaviors/UBehavior.html" title="class in ure.behaviors"><span class="typeNameLink">UBehavior</span></a>
+<li class="circle">ure.actors.behaviors.<a href="../../ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors"><span class="typeNameLink">UBehavior</span></a>
 <ul>
-<li class="circle">ure.behaviors.<a href="../../ure/behaviors/UBehaviorWander.html" title="class in ure.behaviors"><span class="typeNameLink">UBehaviorWander</span></a></li>
+<li class="circle">ure.actors.behaviors.<a href="../../ure/behaviors/UBehaviorWander.html" title="class in ure.actors.behaviors"><span class="typeNameLink">UBehaviorWander</span></a></li>
 </ul>
 </li>
 </ul>
