+++ docs/ure/dagger/AppComponent.html
@@ -177,7 +177,7 @@ public interface <span class="typeNameLabel">AppComponent</span></pre>
 </tr>
 <tr id="i5" class="rowColor">
 <td class="colFirst"><code>void</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="../../ure/dagger/AppComponent.html#inject-ure.behaviors.UBehavior-">inject</a></span>&#8203;(<a href="../../ure/behaviors/UBehavior.html" title="class in ure.behaviors">UBehavior</a>&nbsp;behav)</code></th>
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="../../ure/dagger/AppComponent.html#inject-ure.actors.behaviors.UBehavior-">inject</a></span>&#8203;(<a href="../../ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors">UBehavior</a>&nbsp;behav)</code></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr id="i6" class="altColor">
@@ -348,13 +348,13 @@ public interface <span class="typeNameLabel">AppComponent</span></pre>
 <pre>void&nbsp;inject&#8203;(<a href="../../ure/terrain/TerrainI.html" title="class in ure.terrain">TerrainI</a>&nbsp;terr)</pre>
 </li>
 </ul>
-<a name="inject-ure.behaviors.UBehavior-">
+<a name="inject-ure.actors.behaviors.UBehavior-">
 <!--   -->
 </a>
 <ul class="blockList">
 <li class="blockList">
 <h4>inject</h4>
-<pre>void&nbsp;inject&#8203;(<a href="../../ure/behaviors/UBehavior.html" title="class in ure.behaviors">UBehavior</a>&nbsp;behav)</pre>
+<pre>void&nbsp;inject&#8203;(<a href="../../ure/behaviors/UBehavior.html" title="class in ure.actors.behaviors">UBehavior</a>&nbsp;behav)</pre>
 </li>
 </ul>
 <a name="inject-ure.ui.modals.UModal-">
