+++ src/main/java/ure/actors/UPlayer.java
@@ -2,7 +2,6 @@ package ure.actors;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import ure.areas.UCell;
-import ure.events.PlayerChangedAreaEvent;
 import ure.math.UColor;
 import ure.areas.UArea;
 import ure.ui.ULight;
