+++ src/main/java/ure/actors/behaviors/BehaviorDeserializer.java
@@ -1,4 +1,4 @@
-package ure.behaviors;
+package ure.actors.behaviors;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.ObjectCodec;
