+++ src/main/java/ure/actors/behaviors/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Behaviors are members of NPCs which emit streams of Actions.
+ *
+ */
+package ure.actors.behaviors;
