--- src/main/java/ure/behaviors/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Behaviors are members of NPCs which emit streams of Actions.
- *
- */
-package ure.behaviors;
