+++ src/main/java/ure/sys/dagger/AppModule.java
@@ -8,8 +8,8 @@ import dagger.Provides;
 import ure.areas.LandscaperDeserializer;
 import ure.areas.UCartographer;
 import ure.areas.ULandscaper;
-import ure.behaviors.BehaviorDeserializer;
-import ure.behaviors.UBehavior;
+import ure.actors.behaviors.BehaviorDeserializer;
+import ure.actors.behaviors.UBehavior;
 import ure.sys.UCommander;
 import ure.actors.ActorDeserializer;
 import ure.actors.UActor;
